wordrec: Fix compiler warning
Fix this MS C compiler warning:

wordrec\lm_pain_points.h(102): warning C4244: 'argument':
 conversion from 'double' to 'float', possible loss of data

+++ src/wordrec/lm_pain_points.h
@@ -100,7 +100,7 @@ class LMPainPoints {
                           WERD_RES *word_res);
 
   // Generate a pain point for the blamer.
-  bool GenerateForBlamer(double max_char_wh_ratio, WERD_RES *word_res,
+  bool GenerateForBlamer(float max_char_wh_ratio, WERD_RES *word_res,
                          int col, int row) {
     return GeneratePainPoint(col, row, LM_PPTYPE_BLAMER, 0.0, false,
                              max_char_wh_ratio, word_res);
